+++ Sources/MapboxNavigationCore/Map/Other/NavigationMapIdentifiers.swift
@@ -23,8 +23,6 @@ extension NavigationMapView {
     enum ImageIdentifier {
         static let markerImage = "default_marker"
         static let midpointMarkerImage = "midpoint_marker"
-        static let routeAnnotationLeftHanded = "RouteInfoAnnotationLeftHanded"
-        static let routeAnnotationRightHanded = "RouteInfoAnnotationRightHanded"
         static let trafficSignal = "traffic_signal"
         static let railroadCrossing = "railroad_crossing"
         static let yieldSign = "yield_sign"