+++ Sources/MapboxNavigationCore/Map/Other/UIFont.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+@_spi(MapboxInternal)
+extension UIFont {
+    public class var defaultRouteAnnotationTextFont: UIFont {
+        .systemFont(ofSize: 18)
+    }
+}